+++ docs/HISTORY.rst
@@ -6,7 +6,8 @@ Changelog
 
 - Add uninstall profile
   [cekk]
-
+- Plone 5 compatibility
+  [cekk]
 
 1.3.3 (2015-09-24)
 ------------------